+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/Image.java
@@ -18,8 +18,10 @@ package com.google.cloud.tools.jib.image;
 
 import com.google.cloud.tools.jib.configuration.Port;
 import com.google.common.collect.ImmutableList;
+import java.time.Instant;
 import java.util.List;
 import java.util.Map;
+import javax.annotation.Nullable;
 
 /** Represents an image. */
 public class Image<T extends Layer> {
@@ -30,9 +32,21 @@ public class Image<T extends Layer> {
     private final ImageLayers.Builder<T> imageLayersBuilder = ImageLayers.builder();
     private final ImmutableList.Builder<String> environmentBuilder = ImmutableList.builder();
 
+    @Nullable private Instant created;
     private ImmutableList<String> entrypoint = ImmutableList.of();
     private ImmutableList<String> javaArguments = ImmutableList.of();
     private ImmutableList<Port> exposedPorts = ImmutableList.of();
+
+    /**
+     * Sets the image creation time.
+     *
+     * @param created the creation time
+     * @return this
+     */
+    public Builder<T> setCreated(Instant created) {
+      this.created = created;
+      return this;
+    }
 
     /**
      * Sets the environment with a map from environment variable names to values.
@@ -117,6 +131,7 @@ public class Image<T extends Layer> {
 
     public Image<T> build() {
       return new Image<>(
+          created,
           imageLayersBuilder.build(),
           environmentBuilder.build(),
           ImmutableList.copyOf(entrypoint),
@@ -128,6 +143,9 @@ public class Image<T extends Layer> {
   public static <T extends Layer> Builder<T> builder() {
     return new Builder<>();
   }
+
+  /** The image creation time. */
+  @Nullable private final Instant created;
 
   /** The layers of the image, in the order in which they are applied. */
   private final ImageLayers<T> layers;
@@ -145,16 +163,23 @@ public class Image<T extends Layer> {
   private final ImmutableList<Port> exposedPorts;
 
   private Image(
+      @Nullable Instant created,
       ImageLayers<T> layers,
       ImmutableList<String> environment,
       ImmutableList<String> entrypoint,
       ImmutableList<String> javaArguments,
       ImmutableList<Port> exposedPorts) {
+    this.created = created;
     this.layers = layers;
     this.environmentBuilder = environment;
     this.entrypoint = entrypoint;
     this.javaArguments = javaArguments;
     this.exposedPorts = exposedPorts;
+  }
+
+  @Nullable
+  public Instant getCreated() {
+    return created;
   }
 
   public ImmutableList<String> getEnvironment() {