+++ jib-maven-plugin/CHANGELOG.md
@@ -4,8 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [unreleased]
 
 ### Added
-- Snapshot dependencies are added as their own layer ([#584](https://github.com/GoogleContainerTools/jib/pull/584))
 
+- Snapshot dependencies are added as their own layer ([#584](https://github.com/GoogleContainerTools/jib/pull/584))
 - `jib:buildTar` goal to build an image tarball at `target/jib-image.tar`, which can be loaded into docker using `docker load` ([#514](https://github.com/GoogleContainerTools/jib/issues/514))
 
 - For Docker Hub, also tries registry aliases when getting a credential from the Docker config
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 ### Fixed
+
+- Using multi-byte characters in container configuration ([#626](https://github.com/GoogleContainerTools/jib/issues/626))
 
 ## 0.9.6